// src/database/entities/role-name.enum.ts
export enum RoleName {
  ROLE_SUPER_ADMIN = 'ROLE_SUPER_ADMIN',
  ROLE_ADMIN = 'ROLE_ADMIN',
  ROLE_COMPANY_ADMIN = 'ROLE_COMPANY_ADMIN',
  ROLE_RRHH = 'ROLE_RRHH',
  ROLE_MANAGER = 'ROLE_MANAGER',
  ROLE_USER = 'ROLE_USER',
  ROLE_AUDITOR = 'ROLE_AUDITOR',
  ROLE_WORKFORCE_REPRESENTATIVE = 'ROLE_WORKFORCE_REPRESENTATIVE'
}